+++ index.d.ts
@@ -24,6 +24,8 @@ export interface OneTimeCodeAuthenticateParameters {
   flowType?: string;
   /** Specify where the provider should redirect the user to in your application. If not specified, will be the current location href. Must be a valid redirect url matching what is defined in the application in the Authress Management portal. (Default: **window.location.href**) */
   redirectUrl?: string;
+  /** A list of audiences to add to the JWT in the `aud` claim. This list must be a subset of the audiences defined for the application. To allow additional audiences, see https://authress.io/app/#/manage?focus=applications */
+  audiences?: string[];
   /** Force getting new credentials. (Default: **false** - only get new credentials if none exist.) */
   force?: boolean;
   /** Remove all cookies, LocalStorage, and SessionStorage related data before logging in. In most cases, this helps prevent corrupted browser state from affecting your user's experience. (Default: **true**) */
@@ -43,6 +45,8 @@ export interface AuthenticationParameters {
   flowType?: string;
   /** Specify where the provider should redirect the user to in your application. If not specified, will be the current location href. Must be a valid redirect url matching what is defined in the application in the Authress Management portal. (Default: **window.location.href**) */
   redirectUrl?: string;
+  /** A list of audiences to add to the JWT in the `aud` claim. This list must be a subset of the audiences defined for the application. To allow additional audiences, see https://authress.io/app/#/manage?focus=applications */
+  audiences?: string[];
   /** Overrides the connection specific properties from the Authress Identity Connection to pass to the identity provider */
   connectionProperties?: Record<string, string>;
   /** Force getting new credentials. (Default: **false** - only get new credentials if none exist.) */